chore: extract remote readiness waiters module
Amp-Thread-ID: https://ampcode.com/threads/T-019e9088-f0fe-77ef-bb39-7723f299fe1e
